feat: added new dialogs for managing menu items, rooms, and staff
- Updated client package dependencies and main files
- Added new Dashboard pages and components for managing menu items, rooms, and staff
- Updated sidebar links and dashboard layout
- Added new API handlers for room and staff management
- Removed old image file and added new avatar images to server

+++ server/src/utils/deleteFIle.js
@@ -1,19 +1,21 @@
-import fs from "fs";
-import * as url from "url";
-import path from "path";
+import fs from 'fs';
+import path from 'path';
+import * as url from 'url';
 
 let deleteFile = (fileName) => {
-    const __dirname = url.fileURLToPath(new URL(".", import.meta.url));
+  const __dirname = url.fileURLToPath(new URL('.', import.meta.url));
 
-    // __dirname gives whole folderpath
-    const directoryPath = path.join(__dirname, `../../public/${fileName}`);
+  // __dirname gives whole folderpath
+  const directoryPath = path.join(__dirname, `../../public/${fileName}`);
 
+  if (fs.existsSync(directoryPath)) {
     fs.unlinkSync(directoryPath, (err) => {
-        if (err) {
-            let error = new Error(err.message);
-            throw error;
-        }
+      if (err) {
+        let error = new Error(err.message);
+        throw error;
+      }
     });
+  }
 };
 
 export default deleteFile;
